+++ vortex-array/Cargo.toml
@@ -25,6 +25,7 @@ arrow-schema = { workspace = true }
 enum-iterator = { workspace = true }
 flatbuffers = { workspace = true }
 flexbuffers = { workspace = true }
+futures-util = { workspace = true }
 humansize = { workspace = true }
 itertools = { workspace = true }
 lazy_static = { workspace = true }
@@ -32,6 +33,7 @@ log = { workspace = true }
 num-traits = { workspace = true }
 num_enum = { workspace = true }
 paste = { workspace = true }
+pin-project = { workspace = true }
 rand = { workspace = true }
 vortex-buffer = { path = "../vortex-buffer" }
 vortex-dtype = { path = "../vortex-dtype", features = ["flatbuffers", "serde"] }
@@ -49,6 +51,7 @@ build-vortex = { path = "../build-vortex" }
 
 [dev-dependencies]
 criterion = { workspace = true }
+tokio = { workspace = true }
 
 [[bench]]
 name = "search_sorted"
