+++ vortex-array/src/lib.rs
@@ -8,10 +8,11 @@ mod data;
 pub mod encoding;
 mod flatten;
 mod implementation;
-mod iter;
+pub mod iter;
 mod metadata;
 mod sampling;
 pub mod stats;
+pub mod stream;
 mod tree;
 mod typed;
 pub mod validity;
@@ -26,7 +27,6 @@ pub use context::*;
 pub use data::*;
 pub use flatten::*;
 pub use implementation::*;
-pub use iter::*;
 pub use metadata::*;
 pub use typed::*;
 pub use view::*;
