+++ vortex-ipc/benches/ipc_array_reader_take.rs
@@ -5,10 +5,10 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use futures_util::{pin_mut, TryStreamExt};
 use itertools::Itertools;
 use vortex::array::primitive::PrimitiveArray;
+use vortex::stream::ArrayStreamExt;
 use vortex::{Context, IntoArray};
 use vortex_dtype::Nullability;
 use vortex_dtype::{DType, PType};
-use vortex_ipc::array_stream::ArrayStreamExt;
 use vortex_ipc::io::FuturesVortexRead;
 use vortex_ipc::writer::StreamWriter;
 use vortex_ipc::MessageReader;
